% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopm.R
\name{auto_discovery2}
\alias{auto_discovery2}
\title{Title}
\usage{
auto_discovery2(sample = 5, verbose = TRUE)
}
\arguments{
\item{sample}{Number of samples}

\item{verbose}{\code{TRUE} or \code{FALSE}}
}
\description{
Title
}
