% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distribution.R
\name{chi_square_distribution}
\alias{chi_square_distribution}
\title{Title}
\usage{
chi_square_distribution(n = 1000, output)
}
\arguments{
\item{n}{N Observations}

\item{output}{Path to store plots}
}
\description{
Title
}
