% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_log.R
\name{load_logs}
\alias{load_logs}
\title{Title}
\usage{
load_logs(filenames, column.names, verbose = TRUE)
}
\arguments{
\item{filenames}{Filenames to load}

\item{column.names}{Column names}

\item{verbose}{\code{TRUE} or \code{FALSE}}
}
\description{
Title
}
