% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmopsR-package.R
\docType{package}
\name{pmopsR-package}
\alias{pmopsR}
\alias{pmopsR-package}
\title{pmopsR: R package to support multiple functions for process analysis, mining, simulation and prediction tasks}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
