% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distribution.R
\name{poisson_distribution}
\alias{poisson_distribution}
\title{Title}
\usage{
poisson_distribution(n = 1000, output = ".")
}
\arguments{
\item{n}{N Observations}

\item{output}{Path to store plots}
}
\description{
Title
}
