% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distribution.R
\name{sample_distribution}
\alias{sample_distribution}
\title{Title}
\usage{
sample_distribution(dist.list = c("normal"), n = 1000, output = ".")
}
\arguments{
\item{dist.list}{List of distributions to simulate}

\item{n}{N Observations}

\item{output}{Path to store plots}
}
\description{
Title
}
