% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmopsR-package.R
\docType{package}
\name{pmopsR-package}
\alias{pmopsR}
\alias{pmopsR-package}
\title{pmopsR: Help business analysts to improve the outcomes of their daily process research practices}
\description{
Collection of functions for process analysis, mining, simulation and prediction tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcppc/pmopsR}
  \item Report bugs at \url{https://github.com/jcppc/pmopsR/issues}
}

}
\author{
\strong{Maintainer}: João Caldeira \email{jcppc@iscte-iul.pt} (\href{https://orcid.org/0000-0003-0960-0179}{ORCID})

Authors:
\itemize{
  \item Bárbara Santos \email{Barbara_Santos@iscte-iul.pt} (\href{https://orcid.org/0000-0003-0960-0170}{ORCID})
}

}
\keyword{internal}
